Moved covariance analysis code into analysis subdirectory
