+++ include/industrial_calibration/optimizations/dh_chain_kinematic_calibration.h
@@ -3,7 +3,7 @@
 #include <industrial_calibration/types.h>
 #include <industrial_calibration/dh_chain.h>
 #include <industrial_calibration/optimizations/utils/ceres_math_utilities.h>
-#include <industrial_calibration/optimizations/utils/covariance_analysis.h>
+#include <industrial_calibration/optimizations/analysis/covariance_analysis.h>
 
 namespace industrial_calibration
 {
