+++ include/industrial_calibration/optimizations/extrinsic_hand_eye.h
@@ -17,7 +17,7 @@
 #pragma once
 
 #include <industrial_calibration/types.h>
-#include <industrial_calibration/optimizations/utils/covariance_types.h>
+#include <industrial_calibration/optimizations/analysis/covariance_types.h>
 
 #include <Eigen/Dense>
 #include <vector>
