+++ src/optimizations/analysis/covariance_types.cpp
@@ -1,4 +1,4 @@
-#include <industrial_calibration/optimizations/utils/covariance_types.h>
+#include <industrial_calibration/optimizations/analysis/covariance_types.h>
 
 #include <algorithm>
 #include <cmath>
