+++ src/optimizations/extrinsic_hand_eye.cpp
@@ -2,7 +2,7 @@
 #include <industrial_calibration/cost_functions/extrinsic_hand_eye.h>
 #include <industrial_calibration/optimizations/utils/ceres_math_utilities.h>
 #include <industrial_calibration/types.h>
-#include <industrial_calibration/optimizations/utils/covariance_analysis.h>
+#include <industrial_calibration/optimizations/analysis/covariance_analysis.h>
 #include <industrial_calibration/exceptions.h>
 
 #include <ceres/ceres.h>
