+++ test/include/industrial_calibration_tests/circle_fit.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include <industrial_calibration/types.h>
-#include <industrial_calibration/optimizations/utils/covariance_types.h>
+#include <industrial_calibration/optimizations/analysis/covariance_types.h>
 
 #include <vector>
 #include <Eigen/Dense>
