+++ test/src/circle_fit.cpp
@@ -1,5 +1,5 @@
 #include <industrial_calibration_tests/circle_fit.h>
-#include <industrial_calibration/optimizations/utils/covariance_analysis.h>
+#include <industrial_calibration/optimizations/analysis/covariance_analysis.h>
 
 #include <ceres/ceres.h>
